swagger updated

+++ routes/auth.js
@@ -1,27 +1,32 @@
-const Joi = require("joi");
-const bcrypt = require("bcrypt");
-const _ = require("lodash");
-const { User } = require("../models/user");
-const { Avatar } = require("../models/avatar");
-const mongoose = require("mongoose");
-const express = require("express");
+const Joi = require('joi');
+const bcrypt = require('bcrypt');
+const _ = require('lodash');
+const auth = require('../middleware/auth');
+const { User } = require('../models/user/user');
+const { Avatar } = require('../models/user/avatar');
+const mongoose = require('mongoose');
+const express = require('express');
 const router = express.Router();
-const validateObjectId = require("../middleware/validateObjectId");
+const validateObjectId = require('../middleware/validateObjectId');
 
-router.post("/:id", [validateObjectId], async (req, res) => {
+router.post('/:id', [validateObjectId], async (req, res) => {
   const { error } = validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
-  let user = await User.findOne({ email: req.body.email });
-  if (!user) return res.status(400).send("Invalid email or password.");
+  let user = await User.findOne({ _id: req.params.id });
+  if (!user)
+    return res.status(404).send('The user with the given Id was not found.');
+
+  user = await User.findOne({ email: req.body.email });
+  if (!user) return res.status(400).send('Invalid email or password.');
 
   const validPassword = await bcrypt.compare(req.body.password, user.password);
-  if (!validPassword) return res.status(400).send("Invalid email or password.");
+  if (!validPassword) return res.status(400).send('Invalid email or password.');
 
   await User.findByIdAndRemove(req.params.id);
 
   const token = user.generateAuthToken();
-  res.header("x-auth-token", token).send();
+  res.header('x-auth-token', token).send(token);
 });
 
 /* router.post("/", async (req, res) => {
@@ -38,14 +43,14 @@ router.post("/:id", [validateObjectId], async (req, res) => {
   res.header("x-auth-token", token).send();
 }); */
 
-router.get("/guest", async (req, res) => {
+router.get('/guest', async (req, res) => {
   const date = Date.now().toString();
-  const name = "guest" + date;
-  const email = name + "@mathsnatch.com";
+  const name = 'guest' + date;
+  const email = name + '@mathsnatch.com';
   const password = date;
-  const avatar = "5e4969d45f83cf2170cf7826";
+  const avatar = '5e446b6a9608ec4934a599ef';
 
-  const user = new User({
+  let user = new User({
     name: name,
     email: email,
     password: password,
@@ -74,9 +79,13 @@ router.get("/guest", async (req, res) => {
   res.write("access_token", token);
   res.end(); */
 
-  res
-    .header("x-auth-token", token)
-    .send(_.pick(user, ["_id", "name", "email", "isAdmin", "isGold"]));
+  user = await User.findById(user._id)
+    .select(
+      '-password -__v -finishedCards._id -finishedCards.cards._id -finishedCards.cards.subjects._id -finishedChallenges._id -finishedChallenges.cards._id -finishedChallenges.cards.subjects._id'
+    )
+    .populate('avatar', 'avatarSvg');
+
+  res.header('x-auth-token', token).send(user);
 });
 
 function validate(req) {
